% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_ics.R
\docType{data}
\name{Abbrev_ics}
\alias{Abbrev_ics}
\title{File containing metabolite information.}
\format{
An object of class \code{data.frame} with 248 rows and 8 columns.
}
\usage{
data(Abbrev_ics)
}
\description{
Contains metabolite information such as ID names, abbreviations, KEGG.ID, number of carbons, and Pathways.
Corresponds to data collected using ICS chromatography
}
\keyword{datasets}
