% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev_vanq.R
\docType{data}
\name{Abbrev_vanq}
\alias{Abbrev_vanq}
\title{File containing metabolite information.}
\format{
An object of class \code{data.frame} with 249 rows and 5 columns.
}
\usage{
data(Abbrev_vanq)
}
\description{
Contains metabolite information such as ID names, abbreviations, KEGG.ID, number of carbons, and Pathways.
}
\keyword{datasets}
