% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\docType{data}
\name{Anno}
\alias{Anno}
\title{File containing metabolite information.}
\format{
An object of class \code{data.frame} with 5 rows and 2 columns.
}
\usage{
data(Anno)
}
\description{
Contains metabolite information with full and abbreviated names for raw data table outputs.
}
\keyword{datasets}
