% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sample_colors.R
\name{assign_sample_colors}
\alias{assign_sample_colors}
\title{assign_sample_colors}
\usage{
assign_sample_colors(info)
}
\arguments{
\item{info}{info sheet Condition and Samples columns required}
}
\value{
info sheet with colors column that are assigned to corresponding conditions
}
\description{
assign_sample_colors
}
\examples{
info<-assign_sample_colors(info)

}
