% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{This function produces a bar plot using data from function bar.}
\usage{
bar(metabolites, bar.type, repeats)
}
\arguments{
\item{metabolites}{Metabolic pathway and what type of data to be plotted, for instance: 'glycolysis' for glycolytic metabolites. Individual metabolites can also be entered as vectors.}

\item{bar.type}{The type of data to be plotted. Can be 'tot' (relative amounts), 'iso' (isotopologue data), 'lab' (percent labeled) or 'FC' (fractioanl contribution)}

\item{repeats}{(Only needed in previous versions of ggplot2) Number of conditions to be plotted}
}
\value{
a plot initialized with ggplot
}
\description{
This function produces a bar plot using data from function bar.
}
\author{
Daniel Braas
}
