% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar_plot}
\alias{bar_plot}
\title{This function produces a bar plot using data from function bar.}
\usage{
bar_plot(a, met, Title, x, y, axis.text.x, scales)
}
\arguments{
\item{a}{is the ggplot variable generated with the bar function}

\item{met}{is the data frame that contains what metabolites should be plotted}

\item{Title}{is the title for the plot}

\item{x}{is the title for the x-axis}

\item{y}{is the title for the y-axis}

\item{axis.text.x}{tells ggplot whether or not to label x-axis ticks}

\item{scales}{fixed or free scales}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function produces a bar plot using data from function bar.
}
\author{
Daniel Braas
}
