% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_update_manual.R
\name{bar_plot_update_manual}
\alias{bar_plot_update_manual}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
bar_plot_update_manual(
  a,
  met,
  Title,
  x,
  y,
  axis.text.x,
  scales,
  type = NULL,
  num_cond = NULL,
  index = NULL
)
}
\arguments{
\item{a}{the initial ggplot object}

\item{met}{a data frame containing the relative amounts of the different metabolites in the experiment}

\item{Title}{title}

\item{scales}{parameter passed to ggplot for the facet_wrap vertical axis scale ('free', 'fixed), from ggplot: Are scales shared across all facets (the default, "fixed"), or do they vary across rows ("free_x"), columns ("free_y"), or both rows and columns ("free").}

\item{type}{the type of analysis or quality control being run ('tf', '')}

\item{num_cond}{the number of conditions in the experiment}
}
\value{
a plot of the relative amounts of a specific metabolites that are part of the specified defined pathway (based on the 'metabolites' parameter in bar_update_manual.R)
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline. Called by bar_update_manual. Finalizes the a page of plots of the relative amounts of the different metabolites that are part of the specified defined pathway.
}
\examples{
bar_plot_update_manual(a, met, Title, x, y, axis.text.x, scales='free', num_cond = n, type = "tf")

}
