% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_update_manual.R
\name{bar_update_manual}
\alias{bar_update_manual}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
bar_update_manual(
  metabolites,
  df,
  repeats,
  n,
  type,
  index = NULL,
  title_type,
  only_M0 = NA
)
}
\arguments{
\item{the}{pathway of interest, a data frame containing the relative amounts of the different metabolites in the experiment, the number of conditions in the experiment, and the type of analysis or quality control being run.}
}
\value{
a page of plots of the relative amounts of the different metabolites that are part of the specified defined pathway ('metabolites' parameter)
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
bar_update_manual(metabolites = 'glycolysis', df = data4, n = num_conditions, type = "blank", title_type = "nonpathway")

}
