% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_accucor1.R
\name{call_accucor1}
\alias{call_accucor1}
\title{call_accucor1
Calls Accucor Natural Isotope Correction for single labeled sets (e.g C13, N15, D).
Converts mid_output data to format for Accucor1.
install.packages("accucor")}
\usage{
call_accucor1(mid_output, abbrev, Title, info)
}
\arguments{
\item{mid_output}{"isotopologue data uncorrected.csv".}

\item{abbrev}{The downloaded google sheets Abbrev_New2 sheet. Columns Abb, Neutral.Formula, and KEGG.ID are used in function.}

\item{Title}{Title to use for naming the corrected file}

\item{info}{Sample info sheet, commonly used when running Metabolomics data. Only the info$Condition column is used (make sure this contains factors).}
}
\value{
corr2 NIC corrected version. Not yet in the format for "MIDS corrected output"
}
\description{
call_accucor1
Calls Accucor Natural Isotope Correction for single labeled sets (e.g C13, N15, D).
Converts mid_output data to format for Accucor1.
install.packages("accucor")
}
\examples{
temp<-call_accucor1(mid_output, Abbrev, Title, info)

}
