% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_accucor2.R
\name{call_accucor2}
\alias{call_accucor2}
\title{call_accucor2
#devtools::install_github("wangyujue23/AccuCor2",INSTALL_opts = "--no-multiarch")
Calls Accucor Natural Isotope Correction for DOUBLE labeled sets.
Iso column options are: C13N15-x-x, C13D-x-x, C12 PARENT, C13-, N15-, D-
Converts mid_output data to format for Accucor2.}
\usage{
call_accucor2(mid_output, abbrev, output_dir, Title)
}
\arguments{
\item{mid_output}{Uncorrected Mid output for DOUBLE labeled set.}

\item{abbrev}{The downloaded Google sheets: Abbrev_New2. Columns Abb, Neutral.Formula,Compound, and KEGG.ID are used in function.}

\item{output_dir}{path}

\item{Title}{Title to use for naming the corrected file}
}
\value{
corr3 NIC corrected version. Not yet in the format for "MIDS corrected output"
}
\description{
call_accucor2
#devtools::install_github("wangyujue23/AccuCor2",INSTALL_opts = "--no-multiarch")
Calls Accucor Natural Isotope Correction for DOUBLE labeled sets.
Iso column options are: C13N15-x-x, C13D-x-x, C12 PARENT, C13-, N15-, D-
Converts mid_output data to format for Accucor2.
}
\examples{
temp<-call_accucor2(mid_output, abbrev, output_dir, Title)

}
