% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoAs}
\alias{CoAs}
\title{A character vector of co-enzyme A-related compounds.}
\format{A character vector of length 9.}
\usage{
CoAs
}
\description{
A character vector of co-enzyme A-related compounds.
}
\keyword{datasets}
