% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors}
\alias{colors}
\title{A character vector of commonly used colors.}
\format{A character vector of length 18.}
\usage{
colors
}
\description{
A character vector of commonly used colors.
}
\keyword{datasets}
