% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compound.pathway}
\alias{compound.pathway}
\title{A data frame of 3443 KEGG metabolites and their association with 319 KEGG pathways.}
\format{
A data frame of 5860 rows and 5 variables.
}
\usage{
compound.pathway
}
\description{
A data frame of 3443 KEGG metabolites and their association with 319 KEGG pathways.
}
\keyword{datasets}
