% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compounds}
\alias{compounds}
\title{A data frame of 18042 KEGG compounds.}
\format{
A data frame of 18042 rows and 4 variables.
}
\usage{
compounds
}
\description{
A data frame of 18042 KEGG compounds.
}
\keyword{datasets}
