% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Curr}
\alias{Curr}
\title{A character vector of currency metabolites.}
\format{A character vector of length 16.}
\usage{
Curr
}
\description{
A character vector of currency metabolites.
}
\keyword{datasets}
