% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cys}
\alias{Cys}
\title{A character vector of cysteine metabolism compounds.}
\format{A character vector of length 11.}
\usage{
Cys
}
\description{
A character vector of cysteine metabolism compounds.
}
\keyword{datasets}
