% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cytosine}
\alias{Cytosine}
\title{A character vector of cytosine-related compounds.}
\format{A character vector of length 9.}
\usage{
Cytosine
}
\description{
A character vector of cytosine-related compounds.
}
\keyword{datasets}
