% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotate.R
\name{data_annotate}
\alias{data_annotate}
\title{This function is searching either the KEGG or HMDB databases for potential annotations of ions.}
\usage{
data_annotate(MZ, ppm = 10, adduct, Charge = 1, db = "")
}
\arguments{
\item{MZ}{MZ is the mass to charge (m/z) of the ion(s) of interest}

\item{ppm}{ppm defines the allowed m/z deviation from MZ}

\item{adduct}{Defines the scope of potential ion clusters that will be searched for. This can be set to 'neg' or 'pos' to include all potential ion clusters in a given polarity or can be more specified (e.g. 'M+H')}

\item{Charge}{This variable defines the charge state of the MZ}

\item{db}{This can be used to restrict the search to either KEGG or HMDB. The default is to look in both.}
}
\value{
A data frame with information about the putatively identified feature
}
\description{
This function is searching either the KEGG or HMDB databases for potential annotations of ions.
}
\author{
Daniel Braas
}
