% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_file.R
\docType{data}
\name{element_file}
\alias{element_file}
\title{File containing element information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
data(element_file)
}
\description{
Contains element information for use with IsoCorrectoR.
}
\keyword{datasets}
