% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resp.R
\name{extract_data}
\alias{extract_data}
\title{extract_data}
\usage{
extract_data(data)
}
\arguments{
\item{data}{The ExpeData (".exp" – Sable Systems) file with the spectrograph information for oxygen, carbon dioxide, water vapor flow rate etc.}
}
\description{
Extracts data from SableBase loaded file that is relevant to the user
}
\author{
Daniel Noble – daniel.noble@anu.edu.au
}
