% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FA}
\alias{FA}
\title{A character vector of fatty acid-related compounds.}
\format{A character vector of length 11.}
\usage{
FA
}
\description{
A character vector of fatty acid-related compounds.
}
\keyword{datasets}
