% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
filter_data(data, info, type = "reg")
}
\arguments{
\item{A}{dataframe containing raw metabolomic data. A dataframe containing sample information, specifically condition. An optional parameter 'type' set to 'QC' for regular QC.}
}
\value{
A dataframe of the filtered data. Or if type = 'QC', a vector of the isotopologues to be filtered.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
filter_data(data = data, info = info, type = 'QC')

}
