% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fru}
\alias{Fru}
\title{A character vector of fructose metabolism compounds.}
\format{
A character vector of length 9.
}
\usage{
Fru
}
\description{
A character vector of fructose metabolism compounds.
}
\keyword{datasets}
