% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_medium_plot_func.R
\name{full_medium_plot_func}
\alias{full_medium_plot_func}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
full_medium_plot_func(
  metabolites,
  df,
  n,
  type = NULL,
  index = NULL,
  title_type,
  only_M0 = NA,
  mediumtitle = NULL
)
}
\arguments{
\item{metabolites}{list of metabolites used to run experiment}

\item{df}{amounts3 or amounts}

\item{n}{n = num_conditions}

\item{type}{type = "tf" or other}

\item{index}{index from maven_analysis pre-determined groupings}

\item{title_type}{"nonpathway1", "nonpathway2", or "nonpathway3", or something else}

\item{only_M0}{Not currently being used}

\item{mediumtitle}{mediumtitle="Plots-Relative-To-Fresh"}
}
\value{
a plot of the relative amounts or samples relative to unspent of specific metabolites
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
