% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.pathway}
\alias{gene.pathway}
\title{A data frame of genes associated with 319 KEGG pathways. This data frame is useful for
gene set enrichment analysis if reshaped.}
\format{
A data frame of 25580 rows and 8 variables.
}
\usage{
gene.pathway
}
\description{
A data frame of genes associated with 319 KEGG pathways. This data frame is useful for
gene set enrichment analysis if reshaped.
}
\keyword{datasets}
