% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glycolysis}
\alias{glycolysis}
\title{A character vector of glycolysis compounds.}
\format{
A character vector of length 26.
}
\usage{
glycolysis
}
\description{
A character vector of glycolysis compounds.
}
\keyword{datasets}
