% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Guanine}
\alias{Guanine}
\title{A character vector of guanine metabolism compounds.}
\format{
A character vector of length 8.
}
\usage{
Guanine
}
\description{
A character vector of guanine metabolism compounds.
}
\keyword{datasets}
