% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_auto.R
\name{hand_auto}
\alias{hand_auto}
\alias{ggplot_hand_auto}
\alias{autoplot.hand_auto}
\alias{plot.hand_auto}
\title{Compare hand to auto batch correction}
\usage{
hand_auto(hand, auto)

ggplot_hand_auto(object, ...)

\method{autoplot}{hand_auto}(object, ...)

\method{plot}{hand_auto}(object, ...)
}
\arguments{
\item{hand}{}

\item{auto}{}

\item{object}{object of class `hand_auto`}

\item{...}{parameters ignored}
}
\value{
ggplot object
}
\description{
Compare hand to auto batch correction
}
