% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hex}
\alias{Hex}
\title{A character vector of hexose metabolism compounds.}
\format{
A character vector of length 8.
}
\usage{
Hex
}
\description{
A character vector of hexose metabolism compounds.
}
\keyword{datasets}
