% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_with_subset_blanks.R
\name{info_with_subset_blanks}
\alias{info_with_subset_blanks}
\title{info_with_subset_blanks}
\usage{
info_with_subset_blanks(info)
}
\arguments{
\item{infoA}{data frame that has columns Sample, Run.Order, Condition, etc.}
}
\value{
An info sheet that only contains the two QC-blanks that pertain to this subset.
}
\description{
This function takes a subset sample info sheet and only keeps the two QC-Blanks that were run directly before/ after the samples.
The purpose of this function is to prevent unwanted filtering before calling filter.R,
 that can arise when we include QC-blank values that are not associated with this subset of samples.
}
