% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isobar}
\alias{isobar}
\title{This function takes in an m/z value and will return all possible isobaric compounds
based on the supplied compound data frame supplied in this package.}
\usage{
isobar(MZ = 18.0106, ppm = 5, pol = "positive")
}
\arguments{
\item{MZ}{The m/z value to be searched}

\item{ppm}{The m/z deviation allowed to search the compound library}

\item{pol}{The polarity of the ion to be searched}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function takes in an m/z value and will return all possible isobaric compounds
based on the supplied compound data frame supplied in this package.
}
\author{
Daniel Braas
}
