% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isomer}
\alias{isomer}
\title{This function takes in a chemical formula and will return all isomers found in the
compound library supplied in this package.}
\usage{
isomer(formula = "H2O")
}
\arguments{
\item{formula}{The value for the Type (e.g. H2O for Type Formula)}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function takes in a chemical formula and will return all isomers found in the
compound library supplied in this package.
}
\author{
Daniel Braas
}
