% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isomer_KEGG}
\alias{isomer_KEGG}
\title{This function takes in a chemical formula and will return all possible isomers based on the
KEGG database. Note that the function will pull the data from the KEGG API, thus will require
an internet connection.}
\usage{
isomer_KEGG(formula)
}
\arguments{
\item{formula}{is the chemical formula entered with quotation marks, ie as string}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function takes in a chemical formula and will return all possible isomers based on the
KEGG database. Note that the function will pull the data from the KEGG API, thus will require
an internet connection.
}
\seealso{
\code{\link{keggFind}}
}
