% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG.pathway}
\alias{KEGG.pathway}
\title{A data frame with 514 KEGG pathways and their corresponding IDs.}
\format{A data frame with 514 rows and 2 variables.}
\usage{
KEGG.pathway
}
\description{
A data frame with 514 KEGG pathways and their corresponding IDs.
}
\keyword{datasets}
