\name{make_data_labeled}
\alias{make_data_labeled}
\title{This function takes in a dataframe with the Mass Isotopologue Distributions (MIDs) calculations for partially labeled experiments and outputs a dataframe with percent labeled calculations.}
\usage{
make_data_labeled(DF = data3)
}
\arguments{
\item{DF}{A dataframe with calculated Mass Isotopologue Distribution data.}
}
\description{
This function creates a dataframe with percent labeled calculations.
}
\author{
Gerardo Javier, Daniel Tan, Jordan Uyeki
}


