% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FC.R
\name{make_FC}
\alias{make_FC}
\title{This function produces a data frame with fractional contribution data.}
\usage{
make_FC(DF)
}
\arguments{
\item{DF}{The input data. This should be a data frame that contains MID data generated with the 'make_MID' function.}
}
\value{
A data frame with fractional contribution data.
}
\description{
This function produces a data frame with fractional contribution data.
}
\author{
Daniel Braas
}
