% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap2.R
\name{make_heatmap2}
\alias{make_heatmap2}
\title{Make heatmap}
\usage{
make_heatmap2(
  matrix,
  samples = samples,
  heat.color = normal,
  cluster_samples = TRUE,
  width = NA,
  height = NA,
  ext = NA
)
}
\arguments{
\item{matrix}{Data matrix}

\item{samples}{Sample info sheet data}

\item{heat.color}{Color}

\item{cluster_samples}{Logical; T to cluster}

\item{width}{Plot width}

\item{height}{Plot height}
}
\value{
Heatmap
}
\description{
Make heatmap
}
