% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line_graph.R
\name{make_line_graph}
\alias{make_line_graph}
\title{This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.}
\usage{
make_line_graph(data_frame, iso)
}
\arguments{
\item{data_frame}{The input data. This should be a data frame with Name, Iso, Condition, Time, Exp, Value, Used_ID, KEGG.ID,
Nr.C, Rt and Formula columns. The function assumes that there is a column (Time), which is a factor variable.}

\item{iso}{The isotopologues that are to be plotted.}
}
\value{
A line graph for the specified metabolites and isotopologues.
}
\description{
This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.
}
\author{
Daniel Braas
}
