% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MA_PItbl.R
\name{make_MA_PItbl}
\alias{make_MA_PItbl}
\title{Make MetaboAnalyst compatible peak intensity table CSV}
\usage{
make_MA_PItbl(data_dir, info, Abbrev, abbr = F)
}
\arguments{
\item{data_dir}{Path to network data folder; from pipeline_meta}

\item{info}{Sample info sheet dataframe; from pipeline_meta}

\item{Abbrev}{Abbreviation dataframe; from pipeline_meta}

\item{abbr}{Logical; 'T' to replace names with abbreviations}
}
\value{
Peak intensity CSV in data_dir / network folder
}
\description{
Uses "MavenData" file and reformats to be used with MetaboAnalyst (https://www.metaboanalyst.ca/)
}
