% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{This function produces a matrix for hierarchical clustering or PCA}
\usage{
make_matrix(MS_data, Type, anova = 0.05)
}
\arguments{
\item{MS_data}{The data frame to be used for the matrix. Can be relative amounts, fractional contribution or percent labeled.}

\item{Type}{Which type of data is used. Right now this is not a used variable, but will become important if I add something for isotopologues.}

\item{anova}{A cutoff value: usually a p-value of some sort}
}
\value{

}
\description{
This function produces a matrix for hierarchical clustering or PCA
}
\author{
Daniel Braas
}
