% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MID2.R
\name{make_MID2}
\alias{make_MID2}
\title{This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.}
\usage{
make_MID2(
  DF,
  isotope_correction_remove_small_values_flag = FALSE,
  isotope_correction_remove_small_values_threshold = 0.1
)
}
\arguments{
\item{DF}{The input data. This should be a data frame with Name, Iso, Condition, Exp, Value, Used_ID, KEGG.ID, Nr.C, Rt and Formula columns.}

\item{isotope_correction_remove_small_values_flag}{logical flag for wheter to run the remove small values code}

\item{isotope_correction_remove_small_values_threshold}{threshold used in the remove small values code}
}
\value{
A data frame with MID data.
}
\description{
This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.
}
\author{
Daniel Braas, Thomas Graeber
}
