% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MID_uncorrected.R
\name{make_MID_uncorrected}
\alias{make_MID_uncorrected}
\title{This function produces a data frame with mass isotopologue distribution (MID) data that is NOT corrected for naturally occurring 13C.}
\usage{
make_MID_uncorrected(DF)
}
\arguments{
\item{DF}{The input data. This should be a data frame with Name, Iso, Condition, Exp, Value, Used_ID, KEGG.ID, Nr.C, Rt and Formula columns.}
}
\value{
A data frame with MID data.
}
\description{
This function produces a data frame with mass isotopologue distribution (MID) data that is NOT corrected for naturally occurring 13C.
}
\author{
Daniel Braas, Thomas Graeber
}
