% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mzXML.R
\name{make_mzXML}
\alias{make_mzXML}
\title{This function converts .raw files into .mzXML files for analysis with Maven and XCMS.}
\usage{
make_mzXML(polarity)
}
\arguments{
\item{polarity}{the output polariy. This can be either 'pos', 'neg', 'both', or 'no' for no polarity separation.}
}
\description{
This function converts .raw files into .mzXML files for analysis with Maven and XCMS.
}
\details{
For this function to work, one needs to define where the msconvert.exe file is located eg
D:\\pwiz\\msconvert.exe. This is the location of the program
you will use to convert the MS files. Also, you will need to go into the folder that contains
the .raw files. The function will create directories based on the polarity paramter that is specified.
}
\author{
Daniel Braas
}
