% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PCA.R
\name{make_PCA}
\alias{make_PCA}
\title{This function takes in a data matrix and produces a PCA plot including the top30 loadings.}
\usage{
make_PCA(matrix, a, b)
}
\arguments{
\item{matrix}{The data matrix to be used.}

\item{a}{The PC used for the x-axis.}

\item{b}{The PC used for the y-axis.}
}
\value{
A pdf file with a scree plot, a pair plot showing the first five PCs as well as a PCA plot with PCs a and b as specified in the function call and the corresponding top 30 loadings as bar plot.
}
\description{
This function takes in a data matrix and produces a PCA plot including the top30 loadings.
}
\author{
Daniel Braas
}
