% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PCA2_update.R
\name{make_PCA2_update}
\alias{make_PCA2_update}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
make_PCA2_update(matrix, a = 1, b = 2, cutoff = 0.5)
}
\arguments{
\item{The}{data matrix to be used, the PC to be used for the x axis (default is 1), the PC to be used for the y axis (default is 2), and the cutoff value for minimum Correlation (default is 0.5).}
}
\value{
A pdf file with a scree plot, a pair plot with the first five PCs, and a PCA plot with PCs a and b.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
