% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pheatmap.R
\name{make_pheatmap}
\alias{make_pheatmap}
\title{This function takes in a data matrix and produces an annotated heatmap using the pheatmap package.}
\usage{
make_pheatmap(matrix, samples = samples, heat.color = normal, Norv, Title)
}
\arguments{
\item{matrix}{The data matrix to be used}

\item{samples}{A data frame with information about the Conditions and Cell.Number}

\item{heat.color}{The color scheme to be used for the heatmap.}

\item{Norv}{The internal standard}

\item{Title}{The title to be used. This title will also be part of the file name.}
}
\value{
An annotated heatmap saved as a pdf file.
}
\description{
This function takes in a data matrix and produces an annotated heatmap using the pheatmap package.
}
\author{
Daniel Braas
}
