% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RelAmounts.R
\name{make_RelAmounts}
\alias{make_RelAmounts}
\title{This function produces a data frame with relative metabolite amounts.}
\usage{
make_RelAmounts(DF)
}
\arguments{
\item{DF}{The input data. This should be a data frame with Name, Iso, Condition, Exp, Value, Used_ID, KEGG.ID, Nr.C, Rt and Formula columns.}
}
\value{
A data frame with relative amount information both averaged as well as averaged and normalized to the first condition.
}
\description{
This function produces a data frame with relative metabolite amounts.
}
\author{
Daniel Braas
}
