% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano.R
\name{make_volcano}
\alias{make_volcano}
\title{This function produces a volcano plot using data with two different conditions to be compared.}
\usage{
make_volcano(data, a, b)
}
\arguments{
\item{data}{input data frame}

\item{a}{Numeric value specifying the first condition to evaluate.}

\item{b}{Numeric value specifying the second condition to evaluate.}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function produces a volcano plot using data with two different conditions to be compared.
}
\author{
Daniel Braas
}
