% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapvalues}
\alias{mapvalues}
\title{This function uses the mapvalues function from plyr without the need to load the package.}
\usage{
mapvalues(data, from, to)
}
\arguments{
\item{data}{The data (column) to be modified}

\item{from}{The original value for what is to be renamed}

\item{to}{The value something is mapped to}
}
\value{
A vector (within a data table) with re-mapped values
}
\description{
This function uses the mapvalues function from plyr without the need to load the package.
}
