% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven_MID_Plot.R
\name{Maven_MID_plot}
\alias{Maven_MID_plot}
\title{Maven_MID_plot}
\usage{
Maven_MID_plot(
  metabolites,
  df,
  repeats,
  n,
  type,
  index = NULL,
  title_type,
  only_M0 = NA
)
}
\arguments{
\item{metabolites}{}

\item{df}{}

\item{repeats}{}

\item{n}{number of conditions}

\item{type}{}

\item{index}{}

\item{title_type}{}

\item{only_M0}{}
}
\value{

}
\description{
Maven_MID_plot
}
