\name{maven_processing}
\alias{maven_processing}
\title{Conversion of maven raw file format to tracefinder format
}
\usage{
maven_processing(df)
}
\arguments{
\item{df}{The raw maven data file}

}
\description{
This function takes in a raw maven data file and converts it to tracefinder format
}
\author{
Daniel J Tan
}
