% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavendata_checkup.R
\name{mavendata_checkup}
\alias{mavendata_checkup}
\title{mavendata_checkup}
\usage{
mavendata_checkup(file_path)
}
\arguments{
\item{file_path}{string path of mavendata file}
}
\description{
mavendata_checkup
}
