% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_C.R
\docType{data}
\name{molecule_C}
\alias{molecule_C}
\title{File containing metabolite information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 244 rows and 3 columns.
}
\usage{
data(molecule_C)
}
\description{
Contains molecule information for use with IsoCorrectoR on Carbon Labeled experiments.
}
\keyword{datasets}
