% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_CN.R
\docType{data}
\name{molecule_CN}
\alias{molecule_CN}
\title{File containing metabolite information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 243 rows and 3 columns.
}
\usage{
data(molecule_CN)
}
\description{
Contains molecule information for use with IsoCorrectoR on Carbon-Nitrogen Labeled experiments.
}
\keyword{datasets}
