% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_N.R
\docType{data}
\name{molecule_N}
\alias{molecule_N}
\title{File containing metabolite information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 167 rows and 3 columns.
}
\usage{
data(molecule_N)
}
\description{
Contains molecule information for use with IsoCorrectoR on Nitrogen Labeled experiments.
}
\keyword{datasets}
