% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{MR}
\alias{MR}
\title{MR}
\usage{
MR(v, time, O2_data, CO2_data, method = c("essie", "chris"))
}
\arguments{
\item{v}{The volume of air sample for closed system respirometry. This volume excludes the mass of the animal}

\item{time}{The duration animal was enclosed within the chamber}

\item{O2_data}{The metabR resp data for oxygen}

\item{CO2_data}{The metabR resp data for carbon dioxide}

\item{method}{The method for calculation of metabolic rate using oxygen and carbon dioxide.}
}
\description{
Finds the relevant matching marker and peak data and merges the two data frames together.
}
\author{
Daniel Noble – daniel.noble@anu.edu.au
}
