% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{MRO2}
\alias{MRO2}
\title{MRO2}
\usage{
MRO2(v, time, bp, t, O2_dataChangeChannel, CO2_dataChange)
}
\arguments{
\item{v}{The volume of air sample for closed system respirometry. This volume excludes the mass of the animal}

\item{time}{The duration animal was enclosed within the chamber}

\item{bp}{The barometric pressure in kPa at the time the chamber was closed}

\item{t}{The temperature of the chamber air in degrees Celsius (note this is converted to Kelvins)}

\item{O2_data}{The metabR resp data for oxygen}

\item{CO2_data}{The metabR resp data for carbon dioxide}
}
\description{
Finds the relevant matching marker and peak data and merges the two data frames together.
}
\author{
Daniel Noble – daniel.noble@anu.edu.au

Essie Rodgers – essie.rodgers@anu.edu.au
}
