% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_extract.R
\name{na_rm_extract}
\alias{na_rm_extract}
\title{na_rm_extract}
\usage{
na_rm_extract(data)
}
\arguments{
\item{metabR_data}{The processed data metabR data list from plot_resp}
}
\description{
Removes NA peaks that were detected, caution in case if you forgot to put in marker
}
\author{
Daniel Noble ??? daniel.noble@anu.edu.au and Fonti Kar fonti.kar@gmail.com
}
