% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neurotrans}
\alias{Neurotrans}
\title{A character vector of neurotransmitter compounds.}
\format{A character vector of length 5.}
\usage{
Neurotrans
}
\description{
A character vector of neurotransmitter compounds.
}
\keyword{datasets}
