% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normal}
\alias{normal}
\title{A character vector of colors used in heatmaps.}
\format{
A character vector of length 3.
}
\usage{
normal
}
\description{
A character vector of colors used in heatmaps.
}
\keyword{datasets}
