% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_merge.R
\name{normalize}
\alias{normalize}
\title{Normalize each compound and do log2 transform}
\usage{
normalize(object)
}
\arguments{
\item{object}{data frame}
}
\value{
data frame
}
\description{
Normalize each compound and do log2 transform
}
