% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paths}
\alias{paths}
\title{A list with all information for 319 KEGG pathways.}
\format{
A character vector of length 7.
}
\usage{
paths
}
\description{
A list with all information for 319 KEGG pathways.
}
\keyword{datasets}
