\name{plot_norv}
\alias{plot_norv}
\title{This function takes in a dataframe containing tracefinder data and creates a plot of the Norvaline values in the order that the samples were run.}
\usage{
plot_norv(data = data, info = info_ordered, title = title)
}
\arguments{
\item{Std}{A dataframe containing tracefinder data, including sample Norvaline values}
\item{info}{The information sheet that is ordered in the order that the samples were ran.}
\item{title}{The title of the resulting pdf.}
}
\description{
This function creates a plot of the Norvaline values and returns the resulting dataframe, Norv.
}
\author{
Gerardo Javier, Daniel Tan, Jordan Uyeki
}

