% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resp.R
\name{plot_simple}
\alias{plot_simple}
\title{plot_simple}
\usage{
plot_simple(data, channel, marker, ...)
}
\arguments{
\item{data}{The ExpeData (".exp" – Sable Systems) file with the spectrograph information for oxygen, carbon dioxide, water vapor flow rate etc.}

\item{channel}{The specific channel, or data column from `extract_data` that is relevant for plotting.}

\item{marker}{The marker data that should be plotted.}
}
\description{
Plots the relevant data as a function of time with the marker text plotted as well.
}
\author{
Daniel Noble – daniel.noble@anu.edu.au
}
