% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPP}
\alias{PPP}
\title{A character vector of pentose phosphate metabolism compounds.}
\format{
A character vector of length 14.
}
\usage{
PPP
}
\description{
A character vector of pentose phosphate metabolism compounds.
}
\keyword{datasets}
