% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyrimidine}
\alias{pyrimidine}
\title{A character vector of pyrimidine metabolism compounds.}
\format{
A character vector of length 7.
}
\usage{
pyrimidine
}
\description{
A character vector of pyrimidine metabolism compounds.
}
\keyword{datasets}
