% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_steps.R, R/read_batch_metab.R,
%   R/read_raw_metab.R
\name{qc_steps}
\alias{qc_steps}
\alias{calc_cf}
\alias{replace_missing_ave_cf}
\alias{correct_raw_cf}
\alias{read_batch_metab}
\alias{read_raw_metab}
\title{Title}
\usage{
qc_steps(dirpath, filename, exclude_first = FALSE, Batch_Plate = NULL)

calc_cf(rawobject, exclude_first = FALSE)

replace_missing_ave_cf(qcobject)

correct_raw_cf(rawobject, qcobject)

read_batch_metab(dirpath, filename, sheet = 4)

read_raw_metab(dirpath, filename, sheet = 1, skip = 0, Batch_Plate = NULL)
}
\arguments{
\item{dirpath}{path to directory}

\item{filename}{name of file}

\item{exclude_first}{exclude first QC if `TRUE`}

\item{Batch_Plate}{add `Batch` and `Plate` if not `NULL`.}

\item{rawobject}{}

\item{qcobject}{}

\item{sheet}{sheet number if excel type file}

\item{skip}{number of lines to skip}

\item{object}{object from `calc_cf()`}
}
\value{
data frame
}
\description{
Missing values might be `NA` or `0` values.

Batch corrected values are transposed so steps are different.
}
