% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QE}
\alias{QE}
\title{A character vector of compounds to be used as MID inputs for INCA.}
\format{
A character vector of length 12.
}
\usage{
QE
}
\description{
A character vector of compounds to be used as MID inputs for INCA.
}
\keyword{datasets}
