% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exp.R
\name{read.exp}
\alias{read.exp}
\title{read_resp}
\usage{
read.exp(filename)
}
\arguments{
\item{filename}{Give the path to the .exp file}
}
\description{
read_resp
}
\author{
Thomas Foerster <foerster@sablesys.com>
}
