% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelAmounts.R
\name{RelAmounts}
\alias{RelAmounts}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
RelAmounts(DF, anova = F, type = "NULL", output = F)
}
\arguments{
\item{a}{data frame containing sample data, an optional anova parameter (set to TRUE if ANOVA is to be calculated), an optional type parameter (set to '250K'), and an optional parameter indicating whether or not the dataframe is to be outputted to current working directory.}
}
\value{
a data frame that is in the correct format for the bar_update_manual function to create the Relative Amounts plot.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
