\name{RelAmounts_250K}
\alias{RelAmounts_250K}
\title{This function takes in a data with metabolite data and calculates relative amounts within each metabolite.}
\usage{
RelAmounts_250K(df)
}
\arguments{
\item{df}{A dataframe containing metabolite data.}
}
\description{
This function calculates values, relative to the first condition. If the first condition does not have a value, no modifications to those metabolite's samples are made.
}
\author{
Gerardo Javier, Daniel Tan, Jordan Uyeki
}
