% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_data.R
\name{resp_data}
\alias{resp_data}
\title{resp_data}
\usage{
resp_data(metabR_data)
}
\arguments{
\item{metabR_data}{The processed data metabR data list from plot_resp.}
}
\description{
Finds the relevant matching marker and peak data and merges the two data frames together.
}
\author{
Daniel Noble – daniel.noble@anu.edu.au
}
