% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCA}
\alias{TCA}
\title{A character vector of Tricarboxylic acid (TCA) metabolism compounds.}
\format{A character vector of length 14.}
\usage{
TCA
}
\description{
A character vector of Tricarboxylic acid (TCA) metabolism compounds.
}
\keyword{datasets}
