% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Thymine}
\alias{Thymine}
\title{A character vector of thymine metabolism compounds.}
\format{
A character vector of length 5.
}
\usage{
Thymine
}
\description{
A character vector of thymine metabolism compounds.
}
\keyword{datasets}
