% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_info.R
\name{update_info}
\alias{update_info}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
update_info(info_dir, folder = NULL)
}
\arguments{
\item{The}{path to the original information excel sheet of the analysis.}
}
\value{
An updated excel sheet with sample run orders and the quality control blanks, 250ks, and samples.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
