% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metab.R
\name{write_metab}
\alias{write_metab}
\title{Write Metabolite File}
\usage{
write_metab(object, filename, targeted = TRUE, ...)
}
\arguments{
\item{object}{data frame}

\item{filename}{name of saved file}

\item{targeted}{targeted if `TRUE`}

\item{...}{ignored}
}
\value{
invisible
}
\description{
Write Metabolite File
}
