% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abbrev.R
\docType{data}
\name{Abbrev}
\alias{Abbrev}
\title{File containing metabolite information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 247 rows and 14 columns.
}
\usage{
data(Abbrev)
}
\description{
Contains metabolite information such as ID names, abbreviations, KEGG.ID, number of carbons, and Pathways.
Corresponds to data collected using ICS chromatography
}
\keyword{datasets}
