% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_iso.R
\name{correct_iso}
\alias{correct_iso}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
correct_iso(df, label, correction)
}
\arguments{
\item{the}{isotopomer dataframe, labelling of dataset, and desired molecule(s) for correction}
}
\value{
isotopomer dataframe with a column for corrected values.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
correct_iso(isotopomer_df, "C","C")

}
