% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_medium_plot_func.R
\name{full_medium_plot_func}
\alias{full_medium_plot_func}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
full_medium_plot_func(
  metabolites,
  df,
  repeats,
  n,
  type = NULL,
  index = NULL,
  title_type,
  only_M0 = NA,
  mediumtitle = NULL,
  color_lst = NULL
)
}
\arguments{
\item{metabolites}{list of metabolites used to run experiment}

\item{df}{amounts3 or amounts}

\item{n}{n = num_conditions}

\item{type}{type = "tf" or other}

\item{mediumtitle}{mediumtitle="Plots-Relative-To-Fresh"}

\item{color_lst}{m_ind<-which(!grepl("unspent|[Bb]lank|[Ff]resh|[Cc]ontrol|[Mm]edium|[Mm]edia",unique(amounts$Condition)))   amounts3_colors<- colors1[m_ind]}
}
\value{
a plot of the relative amounts or samples relative to unspent of specific metabolites
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
full_medium_plot_func(unname(unlist(lst)),a, n = num_conditions, type = "tf", title_type = "medium", mediumtitle="Plots-Relative-To-Fresh", color_lst=amounts3_colors)

}
