% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_labeled.R
\name{make_data_labeled}
\alias{make_data_labeled}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
make_data_labeled(DF)
}
\arguments{
\item{The}{dataframe containing the output from the MIDs function.}
}
\value{
A data frame with percent labeled calculations.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
make_data_labeled(DF = data3)

}
