% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap2}
\alias{make_heatmap2}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
make_heatmap2(
  matrix,
  samples = samples,
  heat.color = normal,
  cluster_samples = TRUE,
  width = NA,
  height = NA
)
}
\arguments{
\item{A}{dataframe containing the relative amounts data, a data frame with information on the samples, and the heatmap color scheme. cluster_samples is an optional parameter (default is TRUE) that indicates whether or not the samples are to be clustered. Height and width are optional parameters indicating dimensions of pdf output.}
}
\value{
An annotated heatmap in a pdf file.
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
\examples{
make_heatmap2(matrix = RelA, samples = samples, heat.color = normal, to_cluster = FALSE, width = 8, height = 5)

}
