% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap2.R
\name{make_heatmap2}
\alias{make_heatmap2}
\title{Title}
\usage{
make_heatmap2(
  matrix,
  samples = samples,
  heat.color = normal,
  cluster_samples = TRUE,
  width = NA,
  height = NA
)
}
\arguments{
\item{matrix}{}

\item{samples}{}

\item{heat.color}{}

\item{cluster_samples}{}

\item{width}{}

\item{height}{}
}
\value{

}
\description{
Title
}
