% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_istd.R
\name{make_istd}
\alias{make_istd}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
make_istd(data, remove)
}
\arguments{
\item{A}{dataframe containing the tracefinder data.}
}
\value{
A list of two dataframes, data (the updated data) and Std (the dataframe to be used to make the ISTD plot)
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
