% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven_MID_Plot.R
\name{Maven_MID_plot}
\alias{Maven_MID_plot}
\title{Maven_MID_plot}
\usage{
Maven_MID_plot(
  metabolites,
  df,
  repeats,
  n,
  type,
  index = NULL,
  title_type,
  only_M0 = NA
)
}
\arguments{
\item{df}{data frame such as data3, data4, data_labeled, FC}

\item{repeats}{Not currently being used}

\item{n}{number of conditions}

\item{type}{tf}

\item{index}{index from maven_analysis pre-determined groupings}

\item{title_type}{"nonpathway1", "nonpathway2", or "nonpathway3", or something else}

\item{only_M0}{Not currently being used}
}
\description{
Maven_MID_plot
}
