% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven_processing.R
\name{maven_processing}
\alias{maven_processing}
\title{Functions to assist in the current UCLA Metabolomics Pipeline.}
\usage{
maven_processing(maven_raw)
}
\arguments{
\item{maven_raw}{a metabolomics data frame}
}
\value{
a new data frame with data reformatted
}
\description{
This package contains functions to assist in the current UCLA Metabolomics Quality Control and Analysis Pipeline.
}
