% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resp.R
\name{plot_resp}
\alias{plot_resp}
\title{plot_resp}
\usage{
plot_resp(data, channel, threshold_peak = 0.1, tau = 0.5, ...)
}
\arguments{
\item{data}{The ExpeData (".exp" – Sable Systems) file with the spectrograph information for oxygen, carbon dioxide, water vapor flow rate etc.}
}
\description{
Plots channels with relevant data on oxygen, carbon dioxide, water vapor etc. Finds the local maximum of peaks (carbon dioxide and water vapor only) and spits out the time, marker and maximal value in percentage. It also calculated the percentage change from baseline (uses the median value of entire distribution)
}
\author{
Daniel Noble – daniel.noble@anu.edu.au
}
