\name{update_info}
\alias{update_info}
\title{Update Sample Info Sheet}
\usage{
update_info("/User/SampleType/LabName")
}
\arguments{
\item{update_info}{This parameter is the directory where the info sheet is located. DO NOT have a "/" at the end of the update_info parameter.}
}
\description{
This function takes in the directory of the original info sheet, updates it with all the samples and run orders, and then is written out, overriding the old original info sheet.
}
\author{
Gerardo Javier, Daniel Tan, Jordan Uyeki
}

