% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_info_maven.R
\name{update_info_maven}
\alias{update_info_maven}
\title{Update Maven info sheet}
\usage{
update_info_maven(info_dir)
}
\arguments{
\item{info_dir}{string; path to directory containing original info excel sheet for dataset}
}
\value{
None; writes updated Excel sheet info sheet in specified directory
}
\description{
Updates excel sheet with sample run orders and the quality control blanks, 250ks, pools, and samples. Used for Maven only.
}
